requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
openai>=1.3.5
twilio==8.10.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
alembic==1.13.1
asyncpg==0.29.0
python-dotenv==1.0.0
PyPDF2==3.0.1
scikit-learn==1.3.0
numpy==1.24.3
kafka-python==2.0.2
email-validator>=2.1.0